Add files via upload
I made this game as a final project for my CHE120 project!

It was a creative way to showcase the skills I learned throughout the course, obtain new abilities in Python, and apply them to Pygame. 

The task of this project was to take a well-known game, change some rules and apply some graphics. My group members and I took the game "Guess The Number" that was based on CLUI and updated using graphics from the well known game "Shrek" to make a a GUI-based game. This was done through the use of buttons, text displayed and updated on a screen, while loops, functions for game functionality, and for loops to detect game events.